More progress on starting calls.

+++ ControllingFunction/IControlMessageHandler.cs
@@ -5,6 +5,6 @@ namespace Ropu.ControllingFunction
     public interface IControlMessageHandler
     {
         void Registration(uint userId, ushort rtpPort, ushort floorControlPort, IPEndPoint controlEndpoint);
-        void StartGroupCall(uint userId, uint groupId);
+        void StartGroupCall(uint userId, ushort groupId);
     }
 }
